feat(jobs_manager): move celery app initialization to celery.py refactor(jobs_manager): change import of celery app to relative import
